Add credentials provider in clone method
